Set up annotation processor with test.
This is the start of the rewrite of the fm-sqlite package to use kapt
instead of runtime annotation processing. This will speed up execution
time at runtime at the cost of build time, which is much better in the
long run.

Annotation processing allows us to dynamically generate classes at build
time which will let us generate the menial boilerplate code that usually
goes along with in-code database definitions.

Test: ./gradlew clean test
Issue: #10