Create TableDefinition and remove annotations.
Annotations aren't a good way to store the metadata for a table
definition. Instead of using them, we'll instead simply create a
TableDefinition class that holds the metadata like the constraints and
other create syntax options. We'll still hold a reference to the data
class that both serves as a row and the definition of a table's columns.

Test: ./gradlew clean test