Convert ForeignKey to comply with TableDefinition.
This change also adds comprehensive type safety to ForeignKey by adding
an additional generic parameter which statically binds both columns to
using the same data type. We can also statically guarantee the
ForeignKey child column matches the table it will be added to through
the use of the Constraint generic.

Test: ./gradlew clean test