Remove IndexedColumnGroupClause
Instead of handling both IndexedColumn and IndexedColumnGroup, only
IndexedColumnGroup will be exposed as a vector for handling the
indexed-column clause. This means the IndexedColumn data class is simply
a container class with little logic. The IndexedColumnGroup class acts
like the entire clause (even holding only a single IndexedColumn if need
be). This reduces the code complexity and burden for both users of the
API and the API itself to only need to accept the IndexedColumnGroup
type.

Test: ./gradlew clean test
Issue: none