Rename TableDefinition -> CreateTable
Along with renaming, this change adds the sqlite.statement package which
will house forseeable sql statement like INSERT, SELECT, DELETE, etc. It
made the most sense to package these together as the TableDefintion
class seemed more of a data object that held the fields necessary for a
CREATE TABLE statement. This then allows us to package CREATE TABLE
statements within a (yet to be created) sqlite handler which will
execute the statements.

Test: ./gradlew clean test
Issue: #11